rename schema file
